Support slice mode for SFTP
* Extract SFTP tests to a separate module for readability
* Add an SFTP test where slice mode is activated
* Don't mask exception coming from the FTP backend
* add `retrieveFileStream` in `SFTPClient`
